+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Tue Sep 02 07:41:15 CST 2025
+gradle.version=8.13
