+++ .gradle/config.properties
@@ -1,0 +1,2 @@
+#Tue Sep 02 07:41:04 CST 2025
+java.home=C\:\\Program Files\\Android\\Android Studio\\jbr
