+++ app/src/main/AndroidManifest.xml
@@ -19,21 +19,23 @@
         <meta-data
             android:name="com.google.android.geo.API_KEY"
             android:value="AIzaSyDJ9UnPfDImeu1lO4BP4bwindmb4AiwjgQ" />
+        <meta-data
+            android:name="preloaded_fonts"
+            android:resource="@array/preloaded_fonts" />
 
         <activity
-            android:name=".activity.Home2Activity"
-            >
+            android:name=".activity.HomeActivity">
         </activity>
-        <activity android:name=".activity.MainActivity2"/>
+
         <activity android:name=".activity.MainActivity"
-            android:exported="true" >
+            android:exported="true"
+            android:theme="@style/Theme.InventiApp.Splash">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activity.HomeActivity" />
+
     </application>
 
 </manifest>
