+++ app/src/main/java/com/certicode/inventiapp/activity/MainActivity.kt
@@ -10,6 +10,8 @@ import android.text.style.ForegroundColorSpan
 import android.text.style.StyleSpan
 import android.widget.TextView
 import androidx.core.content.ContextCompat
+import androidx.core.splashscreen.SplashScreen
+import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import com.certicode.inventiapp.R
 import com.certicode.inventiapp.databinding.ActivityHomeBinding
 import com.certicode.inventiapp.databinding.ActivityMainBinding
@@ -19,17 +21,21 @@ class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
+        val splashScreen = installSplashScreen()
         super.onCreate(savedInstanceState)
         binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
         spannableText()
+
+        splashScreen.setKeepOnScreenCondition { false }
+
 
         binding.signInButton.setOnClickListener {
             // This code block is executed when the button is clicked.
 
             // Example: Navigate to a new Activity
             // Replace 'HomeActivity::class.java' with the name of your target Activity.
-            val intent = Intent(this, Home2Activity::class.java)
+            val intent = Intent(this, HomeActivity::class.java)
             startActivity(intent)
         }
     }
