+++ app/src/main/res/drawable/box_background.xml
@@ -1,0 +1,5 @@
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <solid android:color="#E0E0E0" />
+    <corners android:radius="8dp" />
+    <size android:width="60dp" android:height="60dp" />
+</shape>
