+++ app/src/main/res/drawable/splash_logo.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <!-- Background color -->
+    <item android:drawable="@color/white" />
+
+    <!-- Centered logo -->
+    <item
+        android:drawable="@drawable/inventi_logo"
+        android:gravity="center"
+        android:width="90dp"
+        android:height="60dp" />
+
+
+</layer-list>
