+++ app/src/main/res/values/colors.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <color name="black">#FF000000</color>
+    <color name="blue">#1c5196</color>
     <color name="white">#FFFFFFFF</color>
     <color name="bluebg">#281C9D</color>
     <color name="lightbluebg">#281C9D</color>
