+++ app/src/main/res/values/preloaded_fonts.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <array name="preloaded_fonts" translatable="false">
+        <item>@font/poppins</item>
+        <item>@font/poppins_extrabold</item>
+        <item>@font/poppins_semibold</item>
+    </array>
+</resources>
