+++ app/src/main/res/values/strings.xml
@@ -9,4 +9,6 @@
     <string name="add_to_favorites">Add to favorites</string>
     <string name="_5_0">5.0</string>
     <string name="_3_5_km_50_min">3.5 km/50 min</string>
+    <string name="resend_code"><u>Resend code</u></string>
+
 </resources>
