+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-agp = "8.11.1"
+agp = "8.12.2"
 coilSvg = "2.7.0"
 kotlin = "2.0.21"
 coreKtx = "1.16.0"
